+++ aws_annoying/ecs/__init__.py
@@ -1,10 +1,15 @@
+from .check import check_service_task_definition
 from .common import ECSServiceRef
-from .deployment_waiter import ECSDeploymentWaiter
 from .errors import (
     DeploymentFailedError,
     NoRunningDeploymentError,
     ServiceTaskDefinitionAssertionError,
     WaitForDeploymentError,
+)
+from .wait_for import (
+    wait_for_deployment_complete,
+    wait_for_deployment_start,
+    wait_for_service_stability,
 )
 
 __all__ = (
@@ -14,4 +19,8 @@ __all__ = (
     "NoRunningDeploymentError",
     "ServiceTaskDefinitionAssertionError",
     "WaitForDeploymentError",
+    "check_service_task_definition",
+    "wait_for_deployment_complete",
+    "wait_for_deployment_start",
+    "wait_for_service_stability",
 )
